Add block factory class
I've gone back and forth on the best design pattern to use here, but I
think this is a good compromise to start off with.

A new block will be created by calling e.g.
`\Dxw\GovpressBlocks\Factory::create('card')`, which will register all
the fields etc. necessary for that block (and also return the card
instance, although there's nothing you can do with that for now).

I'm also enforcing a kind of singleton here, by causing an exception to be
thrown if you try and new up the same block type twice (which could
cause ACF-related weirdness). But I'm not using an actual singleton
pattern, because the factory-type pattern seems more testable.

This should allow us to add new block types by just writing the
necessary block class, then adding a reference to the `BLOCKS` constant
on the factory.

An actual block will be added in a subsequent commit.