Add basic card block logic
This commit registers the block and fields associated with the card
block.

The accompanying template is just a holder, and will be fleshed out in a
later commit.

The intention is that some of the logic associated with the card block
will be configurable based on the passed in args array (e.g. to load a
custom template, or switch certain fields on/off) in a later commit.